Dashboard 1 0 5 (#5)
* Update groundSettings.xml

* Update resources_en.properties

* Revert "Gps precision update (#1)"

This reverts commit 45a48c0923932c12d74751bde33ccdc067f02a74.

* Delete groundSettings.xml

* Pull in removal of groundSettings.xml and updated slider range calculation (#4)

* Gps precision squash

* Add precision GPS types to telemetry IDs

* Update telemetry parsing to support mixed types

Telemetry parsing now supports mixed type precision for lat/long gps values

* rework bit shifting to be correct

* Update SerialParser.java

* adjust getTelemetry call to handle double instead of float

* Increased the allowed range for settings to 72

* Convert telemetry float values to doubles across the code base

convert telemetry focused float variables and structures to double variables and structures. This probably breaks the graph and the waypoint panel in some way... Testing will bear it out.

* Reworked precision telemetry value parsing logic

...

* Update SerialParser.java

updates for ground settings

Update groundSettings.xml

Update groundSettings.xml

* Revert "Gps precision squash"

This reverts commit 7b997f946494e288931c7e6f4dd0d6439f9e90cf.

* Take abs value of slider range for settings

* Delete groundSettings.xml

Should not be tracked in the repository.

* Update .gitignore

* Update .gitignore

* Add GPS and heading lock labels back into props

+++ src/table/TableFactory.java
@@ -94,7 +94,7 @@ public class TableFactory {
             }
             
             public String getValueAt(int row) {
-                double val = settingList.get(row).getVal();
+                float val = settingList.get(row).getVal();
                 return " "+val;
             }
             
@@ -112,7 +112,7 @@ public class TableFactory {
             
             public void setValueAt(String val, int row) {
             	
-                double newVal = Double.valueOf((String)val);
+                Float newVal = Float.valueOf((String)val);
                 
                 if(settingList.get(row).outsideOfBounds(newVal)) {
                     JFrame mf = new JFrame("Warning");
